more pixiapp changes
