+++ quadratic-client/src/app/gridGL/UI/UIValidations.ts
@@ -73,7 +73,9 @@ export class UIValidations extends Container<SpecialSprite> {
 
   private drawInfiniteRange(range: RefRangeBounds, type: ValidationUIType) {
     const screenRangeRectangle = getRangeRectangleFromCellRefRange(range);
-    const visibleRectangle = sheets.getVisibleRectangle();
+
+    // todo...this needs to be generic for any renderer
+    const visibleRectangle = pixiApp.getVisibleRectangle();
     const intersection = intersects.rectangleClip(screenRangeRectangle, visibleRectangle);
     if (!intersection) {
       return;
